// types/shopping-list-actions.ts
export enum ShoppingListActionType {
  SET_NEW_ITEM_NAME = 'SET_NEW_ITEM_NAME',
  SET_NEW_ITEM_QTY = 'SET_NEW_ITEM_QTY',
  RESET_NEW_ITEM = 'RESET_NEW_ITEM',
  START_EDITING = 'START_EDITING',
  SET_EDIT_NAME = 'SET_EDIT_NAME',
  SET_EDIT_QTY = 'SET_EDIT_QTY',
  SET_EDIT_CATEGORY = 'SET_EDIT_CATEGORY',
  CANCEL_EDITING = 'CANCEL_EDITING',
  SHOW_CATEGORY_MODAL = 'SHOW_CATEGORY_MODAL',
  HIDE_CATEGORY_MODAL = 'HIDE_CATEGORY_MODAL',
  SHOW_DELETE_CONFIRMATION = 'SHOW_DELETE_CONFIRMATION',
  HIDE_DELETE_CONFIRMATION = 'HIDE_DELETE_CONFIRMATION',
}